Remove service definitions on macOS and linux
We want `edgedb server` to be responsive for service files.

This has the downside is that systems which are not managed
by `edgedb server` (i.e. perhaps all production systems) don't have
service files out of the box. This isn't a big downside, though, as
most production systems are managed by ansible or terraform or similar
tools that can generate service files better.