Set ulimit -n to 1024 on centos targets
rpm on centos 7 calls fcntl on every FD up to the max in order to set
CLOEXEC, and the maximum number of open FDs in docker on our runners
was 2**30 - 8 == 1073741816.
(See https://github.com/rpm-software-management/rpm/pull/444).

ulimit can't be configured in a Dockerfile, and there doesn't seem to
be a way to pass argument to `docker build` if you are using the
default github action, so just call ulimit before the big yum install.

There might be a way to configure it on our infra runner side too, I
don't know.