Fix on aarch64 and simplify a bit
All of the conditional logic is totally uneeded, I realize now, since
it is always being run on centos 7.

The issue on aarch64 is that the baseurl in the SCLo sources doesn't
work on arm, since it has `/centos` and not `/altarch`, so we need to
work that out.

Also, switch to the archive.kernel.org mirror.